support multi message packets (#7)
* make sure received packages are always strings, never buffers
* split messages by \r\n
* update orm to 3.8.1 (= sequelize 6.0.0-beta-4)
* update orm to 3.8.2
* npm audit fix
* mark old edit-screenshots for same chart as done

the idea is that in case the network is busy and has lots of edit
screenshots to handle there may be multiple edit-screenshots queued
for the same chart. at some point the render-network will catch up
and perform the edit-screenshots tasks in the order they were created.

once the first of the screenshots has been completed we no can
discard the other screenshot jobs (by marking them done, too) because
the chart has not been edited in the meantime. if the chart would have
been edited after the screenshot was done, new screenshot jobs will
be queued anyway.