support multi message packets (#5)
* json messages are now sent separated by \r\n
* no more need for try catch block here
* save reference to `this` for use in different function
* update orm to 3.8.1 (= sequelize 6.0.0-beta-4)
* increase ping timeout to 60 seconds to compensate for buffered pings
* update orm to 3.8.2
* log more stats about render clients (as perceived from render-server)
* sort jobs by priority (desc) and id (asc)
* load max 1000 jobs at a time
* include total number of jobs in queue in log output
* flag rc ip as dead after trying to restart it 3 times
* don't poll queued jobs based on created_at

since we now limit the number of jobs we load into the JobQueue
on start of the render-server, we need to make sure the remaining
jobs also get processed eventually. Therefor we can't just add
jobs created in last 5 seconds to the JobQueue, as this would
exclude the remaining jobs.

thew new strategy is to work through all remaining jobs, ordered
by priority (highest first) and id (first in, first out).

that means that if there's a large number of low priority screenshot
jobs sitting in the db, the render-network is trying to work through
the entire list now. it would take a while until the list is cleared
but high priority export jobs will still be processed first.

* no need to check health if client is dead already

* log restart attempts and death sentences in action table

* check if a job is still queued before sending it to client

* reset restart attempts after a client has been healthy for a bit

this also delays the restart attempt. instead of restarting after
a single failed health check we're waiting until we have 3
consecutive unhealthy checks.

* only reset restartAttempts if it's not zero

saves us *a lot* of log messages :)