Use node endpoints (#54)
* use new endpoints for node publishing

* fix locator map error

* add static=1 parameter to export url

* give debugger a bit of space

* export separators in footer

* remove check for children vs childNodes length

After a modification in the chart footer markup this check lead
to entire footer blocks being skipped by the export. the byline
block looked like this:

<span class="byline-caption">Chart:</span> <span>Author</span>

which means there's a space between two spans. the space is a
text node so childNodes.length is 3 while children.length is 2.

this is exported completely fine after I removed the check.

* build

* parse new chart blocks regions above-footer and below-footer

above-footer is the region where we now put the chart notes, although
themes may "move" other blocks into this region as well (e.g. to show
the chart source as block element below the notes). Therefor these
regions need to be parsed with walkDom to make sure we catch everything.

the other "new" region is belowFooter which is only below footer in
the DOM markup but is rendered "above" the entire chart in terms of
z-index. This is being used by the new watermark implementation.

* adjust watermark export to new markup

watermarks are now positioned differently than they were before. also
since we're handling watermarks separately from the other nodes (there
even is a processWatermark function) I thought it makes sense to
explicitely include the `.watermark` selector.

* keep at least the 1st-level groups like footer, header etc

the content of the groups will still be flattened

* trim away whitespaces from pseudo element contents

* try to keep text nodes on the page by flooring the y position

also this introduces an error message when text could not be exported
because it is off the page

* build

* exclude belowFooter region in plain mode + parse afterBody region

* keep group node for belowFooter region, too

* build

* add htmlText footer elements comment

* use getAttribute and setAttribute to copy element classes

* adapt parseFooter test to new footer structure

* remove removeActionLinks functionality

* change block classes

* better selector

* update notes selector

* Plugin auth (#80)

* check if user is authorized to use plugin export-pdf

* register available formats

Co-authored-by: ilokhov <ilokhov.dev@gmail.com>
Co-authored-by: Gregor Aisch <gka@users.noreply.github.com>
Co-authored-by: Fabian Gündel <5798652+sto3psl@users.noreply.github.com>