embed emojis as svg (#127)
* embed emojis as svg

* include Twemoji attribution in PDF creator metadata

* fix support for multibyte emojis

* better support for monospace fonts

* don't scale emojis

* remove xml2js dependency

* cache fetch responses to emoji files

* remove twemoji icons from repo

* load twemoji icons from api

* move text splitting into textUtils

* improve validation and set correct content type

* rollup: Rm babel-plugin-transform-async-to-promise

Remove this plugin from export-print, because we ran into a bug that
transpiles 'break' into 'return' when in an async function. See
https://github.com/rpetrich/babel-plugin-transform-async-to-promises/issues/53

We can safely remove this plugin because this code runs in headless
Chrome which supports async/await.

However, still use this plugin in export-pdf and export-svg, because we
need to polyfill async/await support there.

Also update all rollup- and babel-related packages, which requires
a change in rollup.config runtimeHelpers > babelHelpers.

* jsonToPDF: Don't set continued on last text part

Otherwise most of chart text is not positioned correctly.

* jsonToPDF: Fix text after first emoji not rendered

* jsonToPDF: Fix multiline text

The last text part must not have 'continued'.

* jsonToPDF: Fix vertical align

It applies to the whole text not each line.

And refactor the assignment of 'x'. It needs to be reset with each new
line.

* rollup: Put babel helpers into runtime

Like it was before and like it's in datawrapper/datawrapper.

* rollup: Fix IE11 support

* jsonToPDF: Don't crash on emoji network error

Just skip the emoji

* static: Rebuild

Co-authored-by: Jakub Valenta <jakub@datawrapper.de>