feat(datawrapper): show warning when team owner with active subscription assigns new team owner (#337)
* display confirm alert before changing team member's role to owner

* fix(datawrapper) refresh team members after changing a team member's role (#314)

* fire updateUsers event after assigning new role to team member

* calculate isTeamOwner in Members component

After updating the users in the Members component via updateUsers(), the role of the current user needs to be recalculated dynamically as well. Therefore, we don't pass the role property from the php controller (which is static) and define the role property as a computed property in the Members component instead. Accordingly, the property isTeamOwner is recomputed as well.

Co-authored-by: Marten <msigwart@users.noreply.github.com>

* show confirmation modal before changing the team owner

* add role property to svelte store

Certain ui elements change depending on the role of the current user.
Since the role of the current user can change dynamically
(e.g. user sets someone else as team owner --> user becomes admin),
the updated role property needs to be available to several ui components.
Hence, we add it to the store to make it globally available.
The initial value of the role property is set via the php controller.

* remove scope '+product:read' from endpoint

Scope can be removed since this endpoint can only be called by team admins or owners and only returns the products that the team is part of not everything

* add warning to change owner confirmation if current owner has a paid plan (team might lose features)

* create build

Co-authored-by: Marten <msigwart@users.noreply.github.com>